can view cart item in cart queue

+++ app/src/main/java/com/example/rallyapp/database/entities/CartQueueEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "cart_queue")
 data class CartQueueEntity (
-    @PrimaryKey(autoGenerate = true)  @ColumnInfo(name = "id") val id: Long ?,
+    @PrimaryKey(autoGenerate = true)  @ColumnInfo(name = "id") val id: Int ?,
     @ColumnInfo(name = "cart_quantity") var quantity: Int,
     @ColumnInfo(name = "cart_userid") var userId: Int,
     @ColumnInfo(name = "menu_id") var menuId: Int
